photo card 높이 별로 다르게 출력되도록 수정

+++ app/src/main/java/com/example/bookchigibakchigi/data/entity/PhotoCardEntity.kt
@@ -11,7 +11,7 @@ data class PhotoCardEntity(
     @PrimaryKey(autoGenerate = true) val id: Long = 0,
     val imageFileName: String,
     val isbn: String,
-    val width: Long = 0,
-    val height: Long = 0,
+    val width: Int = 0,
+    val height: Int = 0,
     val createdAt: Long, // timestamp
 ) : Parcelable
